A tools application based on windows integrated features.
I created this application mostly to learn and develop my skills using
WPF. This is not a professional tool but you may check the code if you
want to develop something that uses process management or some
application developed in WPF.

The application consists of a main menu that trigger the tools
applications.
You will find few common tools like: Defragmenter, Cleaner, Downloader,
Check Disk, Virus Scanner and a tool for soft "Cryptation" text, that
can be send through SMTP clients after it's configured.

Some of the features may not work as intended, use at own risk as I
didn't have time for testing and fixing all bugs.

You are free to use the code in any way, just to acknowledge the credits
to the developer: Stefancic Roberto Stefan.