This is a tool application based on the windows integrated features.
The main purpose of this project is to expand my knowledge in using C# and WPF, under the .Net platform.

The application has been developed in Visual Studio 2013, requires minimum .net framework 3.5 and provides some useful utilities like: Cleaner, CheckDisk, Defragmenter, Downloader, VirusScaner, Converter and a soft encryption tool for text.
All these tools can be found in the main menu of the application, which has a relative pleasant UI and also offers some fast commands that can be found in one of the XML.

The application code is based mostly on the following concepts:
- process manangement;
- events and delegates;
- reading data from XML;
- algorithms for different tasks.

Along the project, there is also a NSIS installer script that can be used to generate a package with all of the components of the tool.

P.S: This is not a professional tool, use it at own risk. I didn't have the time to test and fix the bugs in it.

Credits: Stefancic Roberto Stefan